Move main to core folder, extract url and protocol
