+++ core/url.py
@@ -1,0 +1,12 @@
+from urllib.parse import urljoin
+from urllib.parse import urlparse
+from urllib.request import pathname2url
+from urllib.request import url2pathname
+
+
+def filename_to_uri(path: str) -> str:
+    return urljoin('file:', pathname2url(path))
+
+
+def uri_to_filename(uri: str) -> str:
+    return url2pathname(urlparse(uri).path)
